# music_app_backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  http.TrackResponse:
    properties:
      album:
        type: string
      artist:
        type: string
      artwork_url:
        type: string
      duration:
        description: Changed to int64
        type: integer
      explicit:
        type: boolean
      external_url:
        type: string
      genre:
        type: string
      id:
        type: string
      popularity:
        type: integer
      preview_url:
        type: string
      provider:
        type: string
      release_date:
        description: Changed to string
        type: string
      title:
        type: string
      track_number:
        type: integer
    type: object
...
  response.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.APIError'
      success:
        type: boolean
    type: object
  response.PaginatedData:
    properties:
      items: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8085
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the backend API for the Music App, providing endpoints for
    authentication, music discovery, playlists, and user library management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Music App Backend API
  version: 1.2.0
paths:
...
  /music/search:
    get:
      description: Searches for tracks by a query string, with optional filters.
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Provider to search (e.g., itunes, spotify)
        in: query
        name: provider
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: size
        type: integer
      - description: Filter by genre
        in: query
        name: genre
        type: string
      - description: Filter by release year
        in: query
        name: year
        type: integer
      - description: Filter by explicit content
        in: query
        name: explicit
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PaginatedData'
                  - properties:
                      tracks:
                        items:
                          $ref: '#/definitions/http.TrackResponse'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      summary: Search for tracks
      tags:
      - Music
  /music/top-charts:
    get:
      description: Retrieves a paginated list of top tracks for a specific country.
      parameters:
      - default: US
        description: Country code (ISO 3166-1 alpha-2)
        in: query
        name: country
        type: string
      - default: itunes
        description: Provider name (e.g., itunes, spotify)
        in: query
        name: provider
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PaginatedData'
                  - properties:
                      tracks:
                        items:
                          $ref: '#/definitions/http.TrackResponse'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      summary: Get top charts
      tags:
      - Music
  /music/tracks/{trackId}:
    get:
      description: Retrieves full details for a single track by its ID and provider.
      parameters:
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      - description: Provider name (e.g., itunes, spotify)
        in: query
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.TrackResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      summary: Get a single track
      tags:
      - Music
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"